Extract App's startup and navigation logic into hooks
App.tsx held all of the startup sequencing (fonts, orientation,
session restore, server-status polling, splash hiding) and the
navigation side-effects (sign-up banner state, post-sign-in redirect,
active-conversation publishing, last-path persistence, notification
navigation) inline. Move the former to app-startup/app-startup.tsx and
the latter to navigation/app-navigation.tsx, leaving App.tsx as wiring
plus the render tree. The pending post-login state lives in the
navigation hook and is passed to the startup hook, which writes it
during session restore. Behavior is unchanged except that
fetchServerStatus no longer closes over the previous status, so the
polling effect no longer needs to restart on status changes.
